Refactors package structure for consent-related classes
